requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
drf-spectacular
gunicorn
dj-database-url
drf-yasg
Pillow